backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
piexif==1.1.3
exifread==3.0.0
python-magic-bin==0.4.14
pytest==7.4.3
pytest-asyncio==0.21.1
